requirements.txt
Flask==2.3.3
psycopg2-binary
python-dotenv
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.4 